refactor(auth): move the reqwest signers into a signers module (pure rename)
First half of splitting tropel-auth into two tiers: the transport-coupled
AuthSigner implementations (Bearer, Basic, ApiKey, OAuth2, SigV4, OAuth1,
Hawk, Digest) move out of lib.rs into signers.rs so the crate root can host
both tiers. Content is byte-identical; the next commit adds the module
glue and the pure flow helpers.
